+++ custom/icds_reports/utils/aggregation_helpers/monolith/agg_child_health.py
@@ -6,7 +6,7 @@ from six.moves import map
 
 from corehq.util.python_compatibility import soft_assert_type_text
 from custom.icds_reports.utils.aggregation_helpers import transform_day_to_month
-from custom.icds_reports.utils.monolith import BaseICDSAggregationHelper
+from custom.icds_reports.utils.aggregation_helpers.monolith import BaseICDSAggregationHelper
 
 
 class AggChildHealthAggregationHelper(BaseICDSAggregationHelper):
