+++ custom/icds_reports/utils/aggregation_helpers/monolith/birth_preparedness_forms.py
@@ -7,7 +7,7 @@ from corehq.apps.userreports.models import StaticDataSourceConfiguration, get_da
 from corehq.apps.userreports.util import get_table_name
 from custom.icds_reports.const import AGG_CCS_RECORD_BP_TABLE
 from custom.icds_reports.utils.aggregation_helpers import month_formatter
-from custom.icds_reports.utils.monolith import BaseICDSAggregationHelper
+from custom.icds_reports.utils.aggregation_helpers.monolith import BaseICDSAggregationHelper
 
 
 class BirthPreparednessFormsAggregationHelper(BaseICDSAggregationHelper):
