+++ custom/icds_reports/utils/aggregation_helpers/monolith/complementary_forms_ccs_record.py
@@ -5,7 +5,7 @@ from dateutil.relativedelta import relativedelta
 
 from custom.icds_reports.const import AGG_CCS_RECORD_CF_TABLE
 from custom.icds_reports.utils.aggregation_helpers import month_formatter
-from custom.icds_reports.utils.monolith import BaseICDSAggregationHelper
+from custom.icds_reports.utils.aggregation_helpers.monolith import BaseICDSAggregationHelper
 
 
 class ComplementaryFormsCcsRecordAggregationHelper(BaseICDSAggregationHelper):
