+++ custom/icds_reports/utils/aggregation_helpers/monolith/ls_awc_visit_form.py
@@ -5,7 +5,7 @@ from dateutil.relativedelta import relativedelta
 
 from custom.icds_reports.const import AGG_LS_AWC_VISIT_TABLE
 from custom.icds_reports.utils.aggregation_helpers import month_formatter
-from custom.icds_reports.utils.monolith import BaseICDSAggregationHelper
+from custom.icds_reports.utils.aggregation_helpers.monolith import BaseICDSAggregationHelper
 
 
 class LSAwcMgtFormAggHelper(BaseICDSAggregationHelper):
